move DumpHandler to own module
